pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fprime-mcp-server"
version = "0.1.0"
description = "F-Prime MCP Server with Microsoft Entra ID Authentication"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "F-Prime Team", email = "fprime@yourcompany.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=0.1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.26.0",
    "msal>=1.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "itsdangerous>=2.1.2",
    "redis>=5.0.0",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
fprime-mcp = "fprime_mcp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/fprime_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]